requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0

# Image processing & metadata extraction
Pillow>=10.0.0
piexif>=1.1.3

# Data handling
pandas>=2.0.0

# Visualization & mapping
matplotlib>=3.7.0
folium>=0.14.0

# Optional: Machine Learning (for future fire detection)
# scikit-learn>=1.3.0
# tensorflow>=2.13.0
# ultralytics>=8.0.0  # YOLOv8

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Development
jupyter>=1.0.0
pytest>=7.4.0
